requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
deepl>=1.17.0